pyproject.toml
[project]
name = "fastapi-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "bcrypt>=4.3.0",
    "cloudinary>=1.44.0",
    "fastapi[standard]>=0.115.12",
    "freezegun>=1.5.2",
    "itsdangerous>=2.2.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "python-slugify>=8.0.4",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "pylint>=3.3.7",
    "ruff>=0.11.12",
]
